CP Base Project
Learn to code a 2D game using Cocos2D (Version 3.4 - Objective C) as a
base and using this set up project. The project will run on iOS, Mac
OS, and Android.

This project is a simple "base" project that will set up all of the
basics for you. This includes:

A start up "Test" Scene
An import file to manage all files needed for importing to other classes
A basic set of resources including a background image, 2 sprites, and
app icons/launch images.

Purpose:

The purpose of this project is for you to download the base project and
follow along with tutorials provided by Stephen McVicker/Neo-Dragon on
Calis Projects Forum and the official YouTube page.